+++ shared/src/map/MapScene.cpp
@@ -104,16 +104,22 @@ std::vector<std::shared_ptr<::PerformanceLoggerInterface>> MapScene::getPerforma
 
 std::vector<std::shared_ptr<LayerInterface>> MapScene::getLayers() {
     std::vector<std::shared_ptr<LayerInterface>> layersList;
-    for (const auto &l : layers) {
-        layersList.emplace_back(l.second);
+    {
+        std::lock_guard<std::recursive_mutex> lock(layersMutex);
+        for (const auto &l: layers) {
+            layersList.emplace_back(l.second);
+        }
     }
     return layersList;
 };
 
 std::vector<std::shared_ptr<IndexedLayerInterface>> MapScene::getLayersIndexed() {
     std::vector<std::shared_ptr<IndexedLayerInterface>> layersList;
-    for (const auto &l : layers) {
-        layersList.emplace_back(std::make_shared<IndexedLayer>(l.first, l.second));
+    {
+        std::lock_guard<std::recursive_mutex> lock(layersMutex);
+        for (const auto &l: layers) {
+            layersList.emplace_back(std::make_shared<IndexedLayer>(l.first, l.second));
+        }
     }
     return layersList;
 }
@@ -476,8 +482,11 @@ void MapScene::drawReadyFrame(const ::RectCoord &bounds, float paddingPc, float 
 
     // for now we only support drawing a ready frame, therefore
     // we disable animations in the layers
-    for (const auto &layer : layers) {
-        layer.second->enableAnimations(false);
+    {
+        std::lock_guard<std::recursive_mutex> lock(layersMutex);
+        for (const auto &layer: layers) {
+            layer.second->enableAnimations(false);
+        }
     }
 
     auto state = LayerReadyState::NOT_READY;
@@ -521,8 +530,11 @@ void MapScene::drawReadyFrame(const ::RectCoord &bounds, float paddingPc, float 
     // re-enable animations if the map scene is used not only for
     // drawReadyFrame
     camera->freeze(false);
-    for (const auto &layer : layers) {
-        layer.second->enableAnimations(true);
+    {
+        std::lock_guard<std::recursive_mutex> lock(layersMutex);
+        for (const auto &layer: layers) {
+            layer.second->enableAnimations(true);
+        }
     }
 }